[VM] add library
