+++ vm/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "vm"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tir = { path = "../tir" }
+pp = { path = "../pp" }
+location = { path = "../location" }
+bigdecimal = "0.3.0"
