+++ vm/src/lib.rs
